Fix: contextのProviderをAuthContextで管理するように修正
