+++ front/src/App.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useState } from "react";
+import React, { useEffect } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
 import "./App.css";
@@ -7,13 +7,18 @@ import Login from "./components/Login/Login";
 import Register from "./components/Register/Register";
 import TopPage from "./components/TopPage/TopPage";
 import { getCurrentUser } from "./lib/api/auth";
-import { User } from "./types";
-import { AuthContext } from "./context/AuthContext";
+import { AuthProvider, useAuthContext } from "./context/AuthContext";
 
-function App() {
-  const [loading, setLoading] = useState<boolean>(true);
-  const [isSignedIn, setIsSignedIn] = useState<boolean>(false);
-  const [currentUser, setCurrentUser] = useState<User | undefined>();
+const App = () => {
+  return (
+    <AuthProvider>
+      <AppContent />
+    </AuthProvider>
+  );
+};
+
+const AppContent = () => {
+  const { setIsSignedIn, setCurrentUser, setLoading } = useAuthContext();
 
   // ログインしているユーザーの情報を取得
   const handleGetCurrentUser = async () => {
@@ -34,30 +39,16 @@ function App() {
     handleGetCurrentUser();
   }, [setCurrentUser]);
 
-  const AuthContextValue = useMemo(
-    () => ({
-      loading,
-      setLoading,
-      isSignedIn,
-      setIsSignedIn,
-      currentUser,
-      setCurrentUser,
-    }),
-    [loading, setLoading, isSignedIn, setIsSignedIn, currentUser, setCurrentUser],
-  );
-
   return (
-    <AuthContext.Provider value={AuthContextValue}>
-      <Router>
-        <Header />
-        <Routes>
-          <Route path="/" element={<TopPage />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/register" element={<Register />} />
-        </Routes>
-      </Router>
-    </AuthContext.Provider>
+    <Router>
+      <Header />
+      <Routes>
+        <Route path="/" element={<TopPage />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/register" element={<Register />} />
+      </Routes>
+    </Router>
   );
-}
+};
 
 export default App;
