+++ front/src/components/Header/Header.tsx
@@ -1,13 +1,13 @@
-import React, { FC, useContext, useState } from "react";
+import React, { FC, useState } from "react";
 import { Link } from "react-router-dom";
 
-import { AuthContext } from "../../context/AuthContext";
+import { useAuthContext } from "../../context/AuthContext";
 import BeforeLoginMenu from "./BeforeLoginMenu";
 import AfterLoginMenu from "./AfterLoginMenu";
 
 const Header: FC = () => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
-  const { isSignedIn } = useContext(AuthContext);
+  const { isSignedIn } = useAuthContext();
 
   const changeMenuState = () => {
     setIsOpen(!isOpen);
