+++ front/src/context/AuthContext.tsx
@@ -1,13 +1,39 @@
-import React, { createContext } from "react";
+import React, { createContext, useContext, useMemo, useState } from "react";
 import { User } from "../types";
 
-export const AuthContext = createContext(
-  {} as {
-    loading: boolean;
-    setLoading: React.Dispatch<React.SetStateAction<boolean>>;
-    isSignedIn: boolean;
-    setIsSignedIn: React.Dispatch<React.SetStateAction<boolean>>;
-    currentUser: User | undefined;
-    setCurrentUser: React.Dispatch<React.SetStateAction<User | undefined>>;
-  },
-);
+type AuthContextType = {
+  loading: boolean;
+  setLoading: React.Dispatch<React.SetStateAction<boolean>>;
+  isSignedIn: boolean;
+  setIsSignedIn: React.Dispatch<React.SetStateAction<boolean>>;
+  currentUser: User | undefined;
+  setCurrentUser: React.Dispatch<React.SetStateAction<User | undefined>>;
+};
+
+type AuthProviderProps = {
+  children: React.ReactNode;
+};
+
+const AuthContext = createContext<AuthContextType>({} as AuthContextType);
+
+export const AuthProvider = ({ children }: AuthProviderProps) => {
+  const [loading, setLoading] = useState<boolean>(true);
+  const [isSignedIn, setIsSignedIn] = useState<boolean>(false);
+  const [currentUser, setCurrentUser] = useState<User | undefined>();
+
+  const AuthContextValue = useMemo(
+    () => ({
+      loading,
+      setLoading,
+      isSignedIn,
+      setIsSignedIn,
+      currentUser,
+      setCurrentUser,
+    }),
+    [loading, setLoading, isSignedIn, setIsSignedIn, currentUser, setCurrentUser],
+  );
+
+  return <AuthContext.Provider value={AuthContextValue}>{children}</AuthContext.Provider>;
+};
+
+export const useAuthContext = () => useContext(AuthContext);
